Fix build warnings

+++ src/singleshottimer.cpp
@@ -51,7 +51,7 @@ void SingleShotTimer::shot(size_t milliseconds, HandlerT handler, bool restart)
     m_handler=std::move(handler);
 
     // in case of successive calls activate timer as fast as possible
-    if (restart || !m_timer.isActive() || m_timer.interval()>milliseconds)
+    if (restart || !m_timer.isActive() || static_cast<size_t>(m_timer.interval())>milliseconds)
     {
         m_timer.setInterval(milliseconds);
         m_timer.start();
